write more sst tests
