+++ src/table/block_cache.rs
@@ -1,0 +1,5 @@
+use std::sync::Arc;
+
+use crate::block::Block;
+
+pub type BlockCache = moka::sync::Cache<(usize, usize), Arc<Block>>;
